+++ jsonschema/src/keywords/minimum.rs
@@ -1,5 +1,5 @@
 use crate::{
-    compilation::{context::CompilationContext, JSONSchema},
+    compilation::context::CompilationContext,
     error::{error, no_error, ErrorIterator, ValidationError},
     keywords::CompilationResult,
     paths::{InstancePath, JSONPointer},
@@ -30,11 +30,10 @@ macro_rules! validate {
         impl Validate for $validator {
             fn validate<'a, 'b>(
                 &self,
-                schema: &'a JSONSchema,
                 instance: &'b Value,
                 instance_path: &InstancePath,
             ) -> ErrorIterator<'b> {
-                if self.is_valid(schema, instance) {
+                if self.is_valid(instance) {
                     no_error()
                 } else {
                     error(ValidationError::minimum(
@@ -46,7 +45,7 @@ macro_rules! validate {
                 }
             }
 
-            fn is_valid(&self, _: &JSONSchema, instance: &Value) -> bool {
+            fn is_valid(&self, instance: &Value) -> bool {
                 if let Value::Number(item) = instance {
                     return if let Some(item) = item.as_u64() {
                         !NumCmp::num_lt(item, self.limit)
@@ -72,7 +71,7 @@ validate!(MinimumU64Validator);
 validate!(MinimumI64Validator);
 
 impl Validate for MinimumF64Validator {
-    fn is_valid(&self, _: &JSONSchema, instance: &Value) -> bool {
+    fn is_valid(&self, instance: &Value) -> bool {
         if let Value::Number(item) = instance {
             return if let Some(item) = item.as_u64() {
                 !NumCmp::num_lt(item, self.limit)
@@ -88,11 +87,10 @@ impl Validate for MinimumF64Validator {
 
     fn validate<'a, 'b>(
         &self,
-        schema: &'a JSONSchema,
         instance: &'b Value,
         instance_path: &InstancePath,
     ) -> ErrorIterator<'b> {
-        if self.is_valid(schema, instance) {
+        if self.is_valid(instance) {
             no_error()
         } else {
             error(ValidationError::minimum(