+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Remove unused private field in `JSONSchema`, that lead to improvement in the compilation performance.
 - Optimize the `multipleOf` implementation, which now can short-circuit in some cases.
 - Add special cases for arrays with 2 and 3 items in the `uniqueItems` keyword implementation.
+- Remove the `schema` argument from all methods of the `Validate` trait.
 
 ## [0.13.2] - 2021-11-04